ai-user-research/requirements.txt
openai>=1.0.0
fastapi
uvicorn
python-dotenv
requests